Create diagrams/representations.py, and migrate amplitude to features.py
This change reflects the new documentation structure and the difference now drawn there between feature creation for diagrams (scalars) and vector representations/vectorizations. These changes are mostly invisible to the user.
