+++ gtda/diagrams/__init__.py
@@ -1,10 +1,12 @@
 """The module :mod:`gtda.diagrams` implements transformers to preprocess
-persistence diagrams or extract features from them."""
+persistence diagrams, extract features from them, or compute pairwise
+distances between diagrams."""
 
 from .preprocessing import ForgetDimension, Scaler, Filtering
-from .distance import PairwiseDistance, Amplitude
-from .features import PersistenceEntropy, BettiCurve, PersistenceLandscape, \
-    HeatKernel, Silhouette, PersistenceImage
+from .distance import PairwiseDistance
+from .features import PersistenceEntropy, Amplitude
+from .representations import BettiCurve, PersistenceLandscape, HeatKernel, \
+    Silhouette, PersistenceImage
 
 __all__ = [
     'ForgetDimension',
